Enabled Synchronous property on Until Successful connector
Enabled Synchronous property on Until Successful connector to fix the following exception:
ERROR 2018-12-07 14:40:23,561 [main] org.mule.module.launcher.application.DefaultMuleApplication: null
java.io.FileNotFoundException: class path resource [properties/globalweather-${mule.envt}.properties] cannot be opened because it does not exist

This process of consuming SOAP web service needs to occur in order to be able to execute rest of the flow. Therefore, it was a better option than creating an In-Memory Object Store to resolve this issue and keeping the connector asynchronous.